core/requirements.txt
# Core Django dependencies
Django==5.0.4
djangorestframework==3.15.1
django-cors-headers==4.3.1
django-filter==24.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2

# Python standard dependencies
pytz==2024.1
sqlparse==0.4.4

# MyCoin Blockchain dependencies
ecdsa==0.18.0
cryptography==41.0.4
base58==2.1.1
